Solutions for all user functionality

+++ authentication/views.py
@@ -1,3 +1,4 @@
+from django.contrib import messages
 from django.shortcuts import render, redirect
 from django.contrib.auth import login, authenticate, logout
 from .models import CustomUser
@@ -23,6 +24,12 @@ def register(request, id=0):
         if id == 0:
             form = RegisterForm()
         else:
+            if not request.user.is_authenticated:
+                messages.info(request, "Log in first!")
+                return redirect("authorise")
+            if not CustomUser.get_by_email(request.user.email).role == 1:
+                messages.info(request, "You don`t have permission!")
+                return redirect("home")
             user = CustomUser.get_by_id(id)
             form = RegisterForm(instance=user)
         context["data"] = form
@@ -55,16 +62,61 @@ def authorise(request):
 
 def log_out(request):
     logout(request)
-    return redirect("home")
+    messages.info(request, "Logged out!")
+    return redirect("authorise")
 
 
 def get_all(request):
+    if not request.user.is_authenticated:
+        messages.info(request, "Log in first!")
+        return redirect("authorise")
+    if not CustomUser.get_by_email(request.user.email).role == 1:
+        messages.info(request, "You don`t have permission!")
+        return redirect("home")
     return render(request, "get_users.html", {"users": CustomUser.get_all()})
 
 
 def delete_user(request, id):
+    if not request.user.is_authenticated:
+        messages.info(request, "Log in first!")
+        return redirect("authorise")
+    if not CustomUser.get_by_email(request.user.email).role == 1:
+        messages.info(request, "You don`t have permission!")
+        return redirect("home")
     CustomUser.delete_by_id(id)
     return redirect("get_users")
+
+
+def user_info(request, id):
+    if not request.user.is_authenticated:
+        messages.info(request, "Log in first!")
+        return redirect("authorise")
+    if not CustomUser.get_by_email(request.user.email).role == 1:
+        messages.info(request, "You don`t have permission!")
+        return redirect("home")
+    return render(request, "get_user_by_id.html", context={"user": CustomUser.get_by_id(id)})
+
+
+def block(request, id):
+    if not request.user.is_authenticated:
+        messages.info(request, "Log in first!")
+        return redirect("authorise")
+    if not CustomUser.get_by_email(request.user.email).role == 1:
+        messages.info(request, "You don`t have permission!")
+        return redirect("home")
+    CustomUser.block(id)
+    return redirect("user_info", id)
+
+
+def change_role(request, id):
+    if not request.user.is_authenticated:
+        messages.info(request, "Log in first!")
+        return redirect("authorise")
+    if not CustomUser.get_by_email(request.user.email).role == 1:
+        messages.info(request, "You don`t have permission!")
+        return redirect("home")
+    CustomUser.change_role(id)
+    return redirect("user_info", id)
 
 
 def home(request):
